start work on task system
added tak ui
