+++ Assets/Scripts/Player/Tutorial.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5b0fd72a700444209bb19a42627eb06a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
